# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: orbeat-api
  description: >-
    orbeat-api — the governed catalog / entitlements / artifacts / audit control-plane API.
    OAuth 2.1 resource server (audience `orbeat-api`); most routes require a bearer access token,
    admin routes additionally require the `orbeat-admin` realm role.
  # Deliberately not a real version: internal/api's openapi_version.go
  # substitutes this exact placeholder for internal/version.Version at
  # serve time (openapi.yaml can't take a linker -ldflags substitution the
  # way a Go binary can — see that file's doc comment). If you ever see this
  # literal value served from GET /openapi.yaml, the substitution broke.
  version: 0.0.0-UNSET
servers:
  - url: /
    description: The orbeat-api server (relative to its deployed origin).
...
security:
  - bearerAuth: []
paths:
...
  /v1/me:
    get:
      tags: [identity]
      summary: Return the caller's token-derived identity.
      description: >-
        Authenticated only (no DB resolve/RBAC) — reflects the validated bearer token's
        subject/email/realm-roles as-is.
      responses:
        "200":
          description: The caller's identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/catalog:
    get:
      tags: [catalog]
      summary: List the MCP servers the caller is entitled to.
      description: >-
        RBAC-filtered by the caller's resolved roles; only `active` servers are returned.
        `allowedTools: null` means all tools on that server are permitted.
      responses:
        "200":
          description: The caller's visible catalog.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/sync/artifacts:
    get:
      tags: [sync]
      summary: List the caller's entitled role-visibility artifacts, rendered to final file content.
      description: >-
        Content is ready-to-write (e.g. subagent memory frontmatter already injected).
        Org-visibility artifacts are never returned here — they ship via the Channel-1 plugin.
      responses:
        "200":
          description: The caller's entitled artifacts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncArtifactsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/sync/config:
    get:
      tags: [sync]
      summary: Advertise client bootstrap config (currently just the gateway resource URL).
      responses:
        "200":
          description: Sync client bootstrap config.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/servers:
    post:
      tags: [servers]
      summary: Create an MCP server catalog entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerInput"
      responses:
        "201":
          description: The created server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMCPServer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    get:
      tags: [servers]
      summary: List MCP server catalog entries for the tenant, keyset-paginated.
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListCursor"
      responses:
        "200":
          description: A page of catalog entries (admin view, includes disabled/draft).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListServersResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/servers/{id}:
    get:
      tags: [servers]
      summary: Fetch one MCP server catalog entry.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: The server.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMCPServer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    put:
      tags: [servers]
      summary: Replace an MCP server catalog entry.
      description: >-
        Requires If-Match (optimistic concurrency, spec §5) — obtain the
        current row_version from GET's `ETag` header or `rowVersion` body
        field first.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerInput"
      responses:
        "200":
          description: The updated server.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMCPServer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [servers]
      summary: Delete an MCP server catalog entry.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/roles:
    post:
      tags: [roles]
      summary: Create a role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: The created role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    get:
      tags: [roles]
      summary: List roles for the tenant, keyset-paginated.
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListCursor"
      responses:
        "200":
          description: A page of roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRolesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/roles/{id}:
    delete:
      tags: [roles]
      summary: Delete a role.
      description: >-
        CASCADES: deletes every entitlement AND artifact entitlement granted to this role
        (ON DELETE CASCADE in the schema — one statement, no application code involved).
        This is intentional — the protection is legibility, not prevention. The response
        reports exactly what was revoked, and the same counts plus the revoked servers'
        and artifacts' names are recorded in the audit log (action `role.delete`) so a
        mistaken deletion is recoverable by inspection (an operator can read what was
        revoked and re-grant it) — it is NOT reversible; there is no undo. Unlike its
        sibling deletes (servers, entitlements), this returns 200 with the revoked counts
        rather than 204: the caller cannot compute them client-side, since the entitlement
        and artifact-entitlement lists are capped by default pagination. No If-Match is
        required — a DELETE is an explicit destructive act, not a replay of stale field
        values (spec §2 "Out — If-Match").
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Deleted; the counts of what the cascade revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleDeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/users/{id}:
    delete:
      tags: [users]
      summary: Delete a user.
      description: >-
        Removes a catalog user row (the identity created by an authenticated request via
        UpsertUser). There is no create/update: a user row is populated only by
        authentication, never by an admin API call. Unlike DELETE /v1/admin/roles/{id},
        this has no cascade to revoke: no table in the schema holds a foreign key to
        users.id (role membership is reconciled from token claims on every request, not
        a stored user-role table), so this returns 200 with the identity it destroyed
        (subject, email, displayName) rather than a count of revoked grants, and the
        same identity is recorded in the audit log (action `user.delete`). Deletion is
        not a ban: a deleted subject who authenticates again is upserted afresh and
        consumes a Community seat again if one is free. No If-Match is required, matching
        the other admin deletes (spec §2 "Out, with reasons, not oversights").
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Deleted; the identity that was removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/entitlements:
    post:
      tags: [entitlements]
      summary: Grant a role access to an MCP server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntitlementInput"
      responses:
        "201":
          description: The created entitlement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entitlement"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    get:
      tags: [entitlements]
      summary: List entitlements for the tenant, keyset-paginated.
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListCursor"
      responses:
        "200":
          description: A page of entitlements (admin view).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEntitlementsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/entitlements/{id}:
    delete:
      tags: [entitlements]
      summary: Revoke an entitlement.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/audit:
    get:
      tags: [audit]
      summary: List audit events, newest first, keyset-paginated.
      parameters:
        - name: limit
          in: query
          description: Page size (default 100, capped at 1000). Must be a positive integer.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: cursor
          in: query
          description: Opaque keyset cursor from a prior response's `nextCursor` (base64url("<unixNano>:<id>")).
          required: false
          schema:
            type: string
      responses:
        "200":
          description: A page of audit events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditPageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /v1/admin/artifacts:
    post:
      tags: [artifacts]
      summary: Create a catalog artifact (skill, subagent, or rule) as a draft.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtifactInput"
      responses:
        "201":
          description: The created artifact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    get:
      tags: [artifacts]
      summary: List artifacts for the tenant, keyset-paginated and slim by default.
      description: >-
        Rows omit the heavy content fields (content, approvedContent, memorySeed,
        approvedMemorySeed — each present as an empty string) unless `include=content`
        is passed.
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListCursor"
        - name: state
          in: query
          description: >-
            Filter to a single approval state. Omit for all states. The four values
            are exhaustive, not aspirational — migration 00006_artifact_approval.sql
            puts a CHECK on artifact.approval_state restricting it to exactly these,
            so a fifth value can never exist in the database. NOTE: the HTTP layer
            itself does not validate ?state, though — an unrecognized value (e.g. a
            typo) is not rejected with 400; it simply matches zero rows in SQL and
            returns an empty 200 page. The enum is safe (it can never under- or
            over-match), just not authoritative for the API's error behavior.
          required: false
          schema:
            type: string
            enum: [draft, pending, approved, rejected]
        - name: include
          in: query
          description: >-
            Pass `content` to restore the heavy fields (content, approvedContent,
            memorySeed, approvedMemorySeed) that list rows omit by default. Any
            other non-empty value is rejected with 400.
          required: false
          schema:
            type: string
            enum: [content]
      responses:
        "200":
          description: A page of artifacts (admin view).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListArtifactsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /v1/admin/marketplace/publish:
    post:
      tags: [marketplace]
      summary: Enqueue a Channel-1 marketplace publish.
      responses:
        "202":
          description: Publish queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplacePublishResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/marketplace/status:
    get:
      tags: [marketplace]
      summary: Report the last publish attempt/success for the tenant's marketplace target.
      responses:
        "200":
          description: The publish status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Missing, expired, or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    MeResponse:
      type: object
      properties:
        subject:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
          nullable: true
      required: [subject, email, roles]
...
    CatalogResponse:
      type: object
      properties:
        servers:
          type: array
          items:
            $ref: "#/components/schemas/MCPServer"
      required: [servers]
...
    SyncArtifactsResponse:
      type: object
      properties:
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/SyncArtifact"
      required: [artifacts]
...